Avoided resolving symlinks in pppd support snippets
The full path to the snippet (for use as a logging reference) was
previously obtained via "readlink -f".

However, this resolved too much when some of the snippets were actually
changed into symlinks, losing the location of the symlink source.

Now the snippet paths are obtained by simple path concatenation,
avoiding the problem.