Moved distfiles out of local overlay subdirectory
The reason is that we are now fully layman-compatible, which means
the local overlay will have little if any signicifance in the future,
as all important overlays will be managed by layman instead.

It also de-couples the distfiles-directory from a specific overlay,
as it might be used as a local mirror for all overlays, including
the ones managed by layman.