Totally new patch approach for squid.conf
The incremental changes in squid.conf from one version to the next
are just too many in order to use normal in-place patching.

Therefore, a new approach has been initiated instead: All non-comments
are copied to the end of the template file, and are then commented
out in the original places.

Only after that the actual patches are applied, which only affect the
new section of the non-commented copies at the end of the modified
template file.

Those transformations themselves are also saved as patches, plus
a patch generating a script file for repeating the initial
transformations whenever upstream provides a new version of the
template configuration file.