First-time initialization now renames .bash_profile into .profile
At least if this is possible.

The problem is that the slim login manager sources /etc/profile and
~/.login, but does not care about ~/.bash_profile or ~/.bash_login.

As both of them normally source ~/.bashrc and this script does all the
important work, it can be a problem if ~/.bash_profile exists and
~/.profile does not. It could lead to ~/.bashrc not being sourced.

Renaming ~/.bash_profile into ~/.profile will solve that problem.