Modified guard selection conditions
Nevertheless, it turns out that applying all those patches on different
platforms is really a major PITA. The problem is that the base files are
completely different, and adding any patches on top of that is actually
a game of fortune.

This was not a problem in the past when only different Gentoo systems
participated in the differentiation process, but it's a total mess when
Debian systems are also included.